requirements.txt
asgiref==3.7.1
certifi==2023.5.7
charset-normalizer==3.1.0
defusedxml==0.7.1
Django==4.2.1
django-extensions==3.2.1
djangorestframework==3.14.0
httpie==3.2.2
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
Pygments==2.15.1
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.1
sqlparse==0.4.4
urllib3==2.0.2
